Conflicts:
	res/drawable-hdpi/ic_multi_sim1.png
	res/drawable-hdpi/ic_multi_sim2.png
	res/drawable-hdpi/ic_multi_sim3.png
	res/drawable-hdpi/ic_multi_sim4.png
	res/drawable-mdpi/ic_multi_sim1.png
	res/drawable-mdpi/ic_multi_sim2.png
	res/drawable-mdpi/ic_multi_sim3.png
	res/drawable-mdpi/ic_multi_sim4.png
	res/drawable-xhdpi/ic_multi_sim1.png
	res/drawable-xhdpi/ic_multi_sim2.png
	res/drawable-xhdpi/ic_multi_sim3.png
	res/drawable-xhdpi/ic_multi_sim4.png
	res/drawable-xxhdpi/ic_multi_sim1.png
	res/drawable-xxhdpi/ic_multi_sim2.png
	res/drawable-xxhdpi/ic_multi_sim3.png
	res/drawable-xxhdpi/ic_multi_sim4.png
	res/drawable-xxxhdpi/ic_multi_sim1.png
	res/drawable-xxxhdpi/ic_multi_sim2.png
	res/drawable-xxxhdpi/ic_multi_sim3.png
	res/drawable-xxxhdpi/ic_multi_sim4.png

+++ src/com/android/phone/msim/MSimCallFeaturesSubSetting.java
@@ -272,7 +272,7 @@ public class MSimCallFeaturesSubSetting extends PreferenceActivity
     private Preference mVoicemailNotificationRingtone;
     private SwitchPreference mVoicemailNotificationVibrate;
 
-    private long mSubId;
+    private int mSubId;
     private int mSlotId;
 
     private class VoiceMailProvider {
